Move declarations to src
