+++ src/declarations/burger.ts
