+++ src/declarations/store.ts
